[#3558] MultilineLiteralBraceCorrector extraction
