+++ lib/rubocop.rb
@@ -141,6 +141,7 @@ require_relative 'rubocop/cop/mixin/unused_argument'
 
 require_relative 'rubocop/cop/correctors/alignment_corrector'
 require_relative 'rubocop/cop/correctors/empty_line_corrector'
+require_relative 'rubocop/cop/correctors/multiline_literal_brace_corrector'
 
 require_relative 'rubocop/cop/bundler/duplicated_gem'
 require_relative 'rubocop/cop/bundler/insecure_protocol_source'
