+++ lib/rubocop/cop/layout/multiline_hash_brace_layout.rb
@@ -112,6 +112,10 @@ module RuboCop
         def on_hash(node)
           check_brace_layout(node)
         end
+
+        def autocorrect(node)
+          MultilineLiteralBraceCorrector.correct(processed_source, node)
+        end
       end
     end
   end
