+++ lib/rubocop/cop/mixin/multiline_literal_brace_layout.rb
@@ -17,28 +17,6 @@ module RuboCop
         check(node)
       end
 
-      def autocorrect(node)
-        if closing_brace_on_same_line?(node)
-          lambda do |corrector|
-            corrector.insert_before(node.loc.end, "\n".freeze)
-          end
-        else
-          # When a comment immediately before the closing brace gets in the way
-          # of an easy correction, the offense is reported but not auto-
-          # corrected. The user must handle the delicate decision of where to
-          # put the comment.
-          return if new_line_needed_before_closing_brace?(node)
-
-          lambda do |corrector|
-            corrector.remove(range_with_surrounding_space(range: node.loc.end,
-                                                          side: :left))
-
-            corrector.insert_after(last_element_range_with_trailing_comma(node),
-                                   node.loc.end.source)
-          end
-        end
-      end
-
       private
 
       # Returns true for the case
@@ -91,33 +69,16 @@ module RuboCop
         end
       end
 
-      def ignored_literal?(node)
-        implicit_literal?(node) || empty_literal?(node) || node.single_line?
+      def empty_literal?(node)
+        children(node).empty?
       end
 
       def implicit_literal?(node)
         !node.loc.begin
       end
 
-      def empty_literal?(node)
-        children(node).empty?
-      end
-
-      def last_element_range_with_trailing_comma(node)
-        trailing_comma_range = last_element_trailing_comma_range(node)
-        if trailing_comma_range
-          children(node).last.source_range.join(trailing_comma_range)
-        else
-          children(node).last.source_range
-        end
-      end
-
-      def last_element_trailing_comma_range(node)
-        range = range_with_surrounding_space(
-          range: children(node).last.source_range,
-          side: :right
-        ).end.resize(1)
-        range.source == ',' ? range : nil
+      def ignored_literal?(node)
+        implicit_literal?(node) || empty_literal?(node) || node.single_line?
       end
 
       def children(node)
