+++ lib/rubocop/cop/util.rb
@@ -282,15 +282,6 @@ module RuboCop
           .sub('Style', 'Styles')
       end
 
-      def display_column(range)
-        line = processed_source.lines[range.line - 1]
-        Unicode::DisplayWidth.of(line[0, range.column])
-      end
-
-      def within?(inner, outer)
-        inner.begin_pos >= outer.begin_pos && inner.end_pos <= outer.end_pos
-      end
-
       private
 
       def directions(side)
